refactor(Problem5): 클래스 분리
