+++ src/main/java/onboarding/problem5/Calculator.java
@@ -1,0 +1,14 @@
+package onboarding.problem5;
+
+public class Calculator {
+    public static boolean isBigger(int money, int baseMoney) {
+        if (money >= baseMoney) {
+            return true;
+        }
+        return false;
+    }
+    public static int getQuotient(int money, int baseMoney) {
+        int quotient = money / baseMoney;
+        return quotient;
+    }
+}
