+++ src/main/java/onboarding/problem5/Constant.java
@@ -1,0 +1,10 @@
+package onboarding.problem5;
+
+import java.util.Collections;
+import java.util.List;
+
+public class Constant {
+    static final int MONEY_MIN = 1;
+    static final int MONEY_MAX = 1_000_000;
+    static final List<Integer> EXCEPTION = Collections.emptyList();
+}
