+++ src/main/java/onboarding/problem5/Controller.java
@@ -1,0 +1,14 @@
+package onboarding.problem5;
+
+
+import java.util.List;
+
+public class Controller {
+    public static List<Integer> getSolution(int money) {
+        if (Validation.check(money)) {
+            List<Integer> result = Money.splitMoney(money);
+            return result;
+        }
+        return Constant.EXCEPTION;
+    }
+}
